% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{finnish_stemmer}
\alias{finnish_stemmer}
\title{Stem Finnish words}
\usage{
finnish_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for Finnish words
}
\examples{
finnish_stemmer(c("taivas"))
}
