% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finnish_stoplist}
\alias{finnish_stoplist}
\title{Finnish stop list}
\format{\link{character} containing the list of words}
\source{
\href{http://members.unine.ch/jacques.savoy/clef/}{Unine website}
}
\usage{
finnish_stoplist
}
\description{
Finnish stop list
}
\examples{
data("finnish_stoplist")
}
\keyword{datasets}
