% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{french_stemmer}
\alias{french_stemmer}
\title{Stem French words}
\usage{
french_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for French words
}
\examples{
french_stemmer(c("tester", "testament", "clients"))
}
