% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{french_stoplist}
\alias{french_stoplist}
\title{French stop list}
\format{\link{character} containing the list of words}
\source{
\href{http://members.unine.ch/jacques.savoy/clef/}{Unine website}
}
\usage{
french_stoplist
}
\description{
French stop list
}
\examples{
data("french_stoplist")
}
\keyword{datasets}
