% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{german_stemmer}
\alias{german_stemmer}
\title{Stem German words}
\usage{
german_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for German words
}
\examples{
german_stemmer(c("kinder"))
}
