% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_stoplist}
\alias{german_stoplist}
\title{German stop list}
\format{\link{character} containing the list of words}
\source{
\href{http://members.unine.ch/jacques.savoy/clef/}{Unine website}
}
\usage{
german_stoplist
}
\description{
German stop list
}
\examples{
data("german_stoplist")
}
\keyword{datasets}
