% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{italian_stemmer}
\alias{italian_stemmer}
\title{Stem Italian words}
\usage{
italian_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for Italian words
}
\examples{
italian_stemmer(c("arrivederci"))
}
