% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_stoplist}
\alias{italian_stoplist}
\title{Italian stop list}
\format{\link{character} containing the list of words}
\source{
\href{http://members.unine.ch/jacques.savoy/clef/}{Unine website}
}
\usage{
italian_stoplist
}
\description{
Italian stop list
}
\examples{
data("italian_stoplist")
}
\keyword{datasets}
