% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{portuguese_stemmer}
\alias{portuguese_stemmer}
\title{Stem Portuguese words}
\usage{
portuguese_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for Portuguese words
}
\examples{
portuguese_stemmer(c("adeus"))
}
