% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spanish_stemmer}
\alias{spanish_stemmer}
\title{Stem Spanish words}
\usage{
spanish_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for Spanish words
}
\examples{
spanish_stemmer(c("perros"))
}
