% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spanish_stoplist}
\alias{spanish_stoplist}
\title{Spanish stop list}
\format{\link{character} containing the list of words}
\source{
\href{http://members.unine.ch/jacques.savoy/clef/}{Unine website}
}
\usage{
spanish_stoplist
}
\description{
Spanish stop list
}
\examples{
data("spanish_stoplist")
}
\keyword{datasets}
