% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swedish_stemmer}
\alias{swedish_stemmer}
\title{Stem Swedish words}
\usage{
swedish_stemmer(words)
}
\arguments{
\item{words}{a \link{character} containing the original words.}
}
\value{
\link{character} with stemmed words.
}
\description{
Stemmer for Swedish words
}
\examples{
swedish_stemmer(c("stiga"))
}
