% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swedish_stoplist}
\alias{swedish_stoplist}
\title{Swedish stop list}
\format{\link{character} containing the list of words}
\source{
\href{http://members.unine.ch/jacques.savoy/clef/}{Unine website}
}
\usage{
swedish_stoplist
}
\description{
Swedish stop list
}
\examples{
data("swedish_stoplist")
}
\keyword{datasets}
