% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{unine}
\alias{unine}
\alias{unine-package}
\title{unine: Unine Light Stemmer for 7 Languages}
\description{
Implementation of "light" stemmers for French, German, Italian, Spanish, Portuguese, Finnish, Swedish. 
  They are based on the same work as the "light" stemmers found in SolR or ElasticSearch. 
  A "light" stemmer consist in removing inflections only for noun and adjectives. 
  Indexing verbs for Italian, French or German is not of primary importance compared to nouns and adjectives. 
  The stemming procedure for French is described in (Savoy, 1999).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/unine}
  \item \url{http://members.unine.ch/jacques.savoy/clef/}
  \item Report bugs at \url{https://github.com/pommedeterresautee/unine/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr} [copyright holder]

Other contributors:
\itemize{
  \item Jacques Savoy \email{Jacques.Savoy@unine.ch} [copyright holder]
}

}
