% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{unine}
\alias{unine}
\alias{unine-package}
\title{unine: Unine Light Stemmer}
\description{
Implementation of "light" stemmers for French, German, Italian, Spanish, Portuguese, Finnish, Swedish. 
  They are based on the same work as the "light" stemmers found in 'SolR' <https://lucene.apache.org/solr/> or 'ElasticSearch' <https://www.elastic.co/fr/products/elasticsearch>. 
  A "light" stemmer consists in removing inflections only for noun and adjectives. 
  Indexing verbs for these languages is not of primary importance compared to nouns and adjectives. 
  The stemming procedure for French is described in (Savoy, 1999) <doi:10.1002/(SICI)1097-4571(1999)50:10%3C944::AID-ASI9%3E3.3.CO;2-H>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pommedeterresautee/unine}
  \item \url{https://pommedeterresautee.github.io/unine/}
  \item \url{http://members.unine.ch/jacques.savoy/clef/}
  \item Report bugs at \url{https://github.com/pommedeterresautee/unine/issues}
}

}
\author{
\strong{Maintainer}: Michaël Benesty \email{michael@benesty.fr} [copyright holder]

Other contributors:
\itemize{
  \item Jacques Savoy \email{Jacques.Savoy@unine.ch} [copyright holder]
}

}
